+++ test/watermark_widget_test.dart
@@ -3,7 +3,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:watermark_widget/watermark_widget.dart';
 
 void main() {
-  testWidgets('WatermarkWidget renders child widget', (WidgetTester tester) async {
+  testWidgets('WatermarkWidget renders child widget',
+      (WidgetTester tester) async {
     await tester.pumpWidget(
       MaterialApp(
         home: WatermarkWidget(
@@ -19,7 +20,8 @@ void main() {
     expect(find.text('Child Widget'), findsOneWidget);
   });
 
-  testWidgets('WatermarkWidget renders at least one CustomPaint', (WidgetTester tester) async {
+  testWidgets('WatermarkWidget renders at least one CustomPaint',
+      (WidgetTester tester) async {
     await tester.pumpWidget(
       MaterialApp(
         home: WatermarkWidget(
@@ -36,7 +38,8 @@ void main() {
     expect(find.byType(CustomPaint), findsAtLeastNWidgets(1));
   });
 
-  testWidgets('WatermarkWidget applies correct structure', (WidgetTester tester) async {
+  testWidgets('WatermarkWidget applies correct structure',
+      (WidgetTester tester) async {
     await tester.pumpWidget(
       MaterialApp(
         home: WatermarkWidget(
@@ -56,8 +59,8 @@ void main() {
     expect(find.byType(CustomPaint), findsAtLeastNWidgets(1));
   });
 
-
-  testWidgets('WatermarkWidget updates when properties change', (WidgetTester tester) async {
+  testWidgets('WatermarkWidget updates when properties change',
+      (WidgetTester tester) async {
     await tester.pumpWidget(
       MaterialApp(
         home: WatermarkWidget(
@@ -73,7 +76,8 @@ void main() {
       ),
     );
 
-    final initialCustomPaints = tester.widgetList<CustomPaint>(find.byType(CustomPaint)).toList();
+    final initialCustomPaints =
+        tester.widgetList<CustomPaint>(find.byType(CustomPaint)).toList();
 
     await tester.pumpWidget(
       MaterialApp(
@@ -90,17 +94,16 @@ void main() {
       ),
     );
 
-    final updatedCustomPaints = tester.widgetList<CustomPaint>(find.byType(CustomPaint)).toList();
+    final updatedCustomPaints =
+        tester.widgetList<CustomPaint>(find.byType(CustomPaint)).toList();
 
     // Verify that the number of CustomPaint widgets hasn't changed
     expect(updatedCustomPaints.length, equals(initialCustomPaints.length));
 
     // Verify that at least one CustomPaint widget has been updated
     expect(
-        updatedCustomPaints.any((updated) =>
-            initialCustomPaints.every((initial) => initial.painter != updated.painter)
-        ),
-        isTrue
-    );
+        updatedCustomPaints.any((updated) => initialCustomPaints
+            .every((initial) => initial.painter != updated.painter)),
+        isTrue);
   });
 }